+++ mdsf/src/testing/mod.rs
@@ -3,3 +3,5 @@ pub mod tools;
 pub const DEFAULT_TEST_FORMATTER_TIMEOUT: u64 = 0;
 
 pub const DEFAULT_TEST_DEBUG_ENABLED: bool = true;
+
+pub const DEFAULT_ERROR_ON_TOOL_MISSING: bool = true;