+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file. Dates are d
 
 #### [Unreleased](https://github.com/hougesen/mdsf/compare/v0.8.5...HEAD)
 
+- fix: error if config could not be loaded [`#947`](https://github.com/hougesen/mdsf/pull/947)
 - test(tools): validate dfmt [`#946`](https://github.com/hougesen/mdsf/pull/946)
 - test(tools): validate csslint [`#945`](https://github.com/hougesen/mdsf/pull/945)
 - test(tools): validate snakefmt [`#944`](https://github.com/hougesen/mdsf/pull/944)